moved queuepage to dedicated folder and updated delete method
